refactor(benchmarking): isolate run implementation imports
