+++ benchmarking/curator_benchmarking/commands/run_impl.py
@@ -1,0 +1,551 @@
+# Copyright (c) 2025, NVIDIA CORPORATION.  All rights reserved.
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+"""Benchmark runner implementation.
+
+The public `run` command keeps argument parsing in `run.py` so minimal host
+environments can import the CLI without Curator/runtime dependencies installed.
+Once execution reaches this module, we are in the environment that is expected
+to run benchmarks, so normal top-level imports are preferred over lazy globals.
+"""
+
+from __future__ import annotations
+
+import json
+import os
+import pickle
+import shutil
+import time
+import traceback
+from collections.abc import Mapping
+from contextlib import nullcontext
+from pathlib import Path
+from typing import TYPE_CHECKING, Any
+
+from loguru import logger
+from runner.env_capture import dump_env
+from runner.gpu_stats_recorder import GPUStatsRecorder
+from runner.path_resolver import set_path_mode
+from runner.process import run_command_with_timeout
+from runner.ray_cluster import (
+    get_ray_cluster_data,
+    setup_ray_cluster_and_env,
+    teardown_ray_cluster_and_env,
+)
+from runner.session import Session
+from runner.sinks.sink import call_sink_hook, initialize_sinks
+from runner.utils import (
+    assert_valid_config_dict,
+    find_result,
+    get_gpu_stats,
+    get_obj_for_json,
+    log_gpu_stats,
+    merge_config_files,
+    remove_disabled_blocks,
+    resolve_env_vars,
+)
+from runner.viewer_url import resolve_viewer_url
+
+if TYPE_CHECKING:
+    from argparse import Namespace
+
+    from runner.datasets import DatasetResolver
+    from runner.entry import Entry
+    from runner.path_resolver import PathResolver
+
+from nemo_curator.pipeline.workflow import WorkflowRunResult
+from nemo_curator.tasks.utils import TaskPerfUtils
+from nemo_curator.utils.file_utils import create_or_overwrite_dir
+
+# Benchmark runs always retain Ray Data scheduler diagnostics for post-run analysis.
+os.environ.setdefault("NEMO_CURATOR_RAY_DATA_DIAGNOSTICS", "1")
+
+
+def ensure_dir(dir_path: Path) -> None:
+    """Ensure dir_path and parents exists, creating them if necessary."""
+    dir_path.mkdir(parents=True, exist_ok=True)
+
+
+def get_entry_script_persisted_data(session_entry_path: Path) -> dict[str, Any]:
+    """Read the files that are expected to be generated by the individual benchmark scripts."""
+    params_json = session_entry_path / "params.json"
+    if not params_json.exists():
+        logger.warning(f"Params JSON file not found at {params_json}")
+        script_params = {}
+    else:
+        with open(params_json) as f:
+            script_params = json.load(f)
+
+    metrics_json = session_entry_path / "metrics.json"
+    if not metrics_json.exists():
+        logger.warning(f"Metrics JSON file not found at {metrics_json}")
+        script_metrics = {}
+    else:
+        with open(metrics_json) as f:
+            script_metrics = json.load(f)
+
+    tasks_pkl = session_entry_path / "tasks.pkl"
+    if not tasks_pkl.exists():
+        logger.warning(f"Tasks pickle file not found at {tasks_pkl}")
+        script_tasks = []
+    else:
+        with open(tasks_pkl, "rb") as f:
+            script_tasks = pickle.load(f)  # noqa: S301
+        if isinstance(script_tasks, (list, WorkflowRunResult, Mapping)):
+            script_metrics.update(TaskPerfUtils.aggregate_task_metrics(script_tasks, prefix="task"))
+        else:
+            msg = f"Invalid tasks type loaded from {tasks_pkl}: {type(script_tasks)}"
+            raise TypeError(msg)
+
+    return {"params": script_params, "metrics": script_metrics}
+
+
+def check_requirements_update_results(  # noqa: C901, PLR0912
+    result_data: dict[str, Any],
+    requirements: dict[str, Any],
+) -> bool:
+    """
+    Check if the benchmark meets the requirements. Creates a new "requirements" key in the result_data
+    dictionary with the results of the requirements checks.
+    Returns True if the benchmark meets the requirements, False otherwise.
+    """
+    meets_requirements = True
+    requirements_data = {}
+
+    for metric_name, requirement_dict in requirements.items():
+        reason_not_met = None
+        actual_value = find_result(result_data, metric_name)
+        if actual_value is None:
+            reason_not_met = f"{metric_name} not found in metrics"
+        else:
+            # Already ensured to not have exact and min/max together in Entry
+            has_exact = "exact_value" in requirement_dict
+            has_min = "min_value" in requirement_dict
+            has_max = "max_value" in requirement_dict
+            if has_exact:
+                exact_value = requirement_dict["exact_value"]
+                if actual_value != exact_value:
+                    reason_not_met = f"{metric_name} != {exact_value}"
+            else:
+                if has_min:
+                    min_value = requirement_dict["min_value"]
+                    if actual_value < min_value:
+                        reason_not_met = f"{metric_name} < {min_value}"
+                if has_max:
+                    max_value = requirement_dict["max_value"]
+                    if actual_value > max_value:
+                        reason_not_met = f"{metric_name} > {max_value}"
+                if not has_min and not has_max:
+                    reason_not_met = f"No min or max value specified for {metric_name}"
+
+        # Update the requirements_data dictionary with the result of the requirements check
+        meets_requirements &= reason_not_met is None
+        if reason_not_met is None:
+            logger.debug(f"\t\t✅ Requirement for {metric_name} was met")
+        else:
+            requirements_data[metric_name] = reason_not_met
+            logger.error(f"\t\t❌ Requirement for {metric_name} was not met: {reason_not_met}")
+
+    result_data["requirements_not_met"] = requirements_data
+    return meets_requirements
+
+
+def run_data_setup_entry(
+    setup_entry: Entry,
+    path_resolver: PathResolver,
+    dataset_resolver: DatasetResolver,
+    setup_root_path: Path,
+) -> bool:
+    """Run one data-prep command before benchmark entries."""
+    setup_path = (setup_root_path / setup_entry.name).absolute()
+    logs_path = setup_path / "logs"
+    stdouterr_path = logs_path / "stdouterr.log"
+    create_or_overwrite_dir(setup_path)
+    ensure_dir(logs_path)
+
+    cmd = setup_entry.get_command_to_run(setup_path, path_resolver, dataset_resolver)
+    logger.info(f"🔧 Running data setup {setup_entry.name}")
+    logger.info(f"\tRunning command {' '.join(cmd) if isinstance(cmd, list) else cmd}")
+    started_exec = time.time()
+    result_data: dict[str, Any] = {
+        "name": setup_entry.name,
+        "success": False,
+        "cmd": cmd,
+        "exec_started_at": started_exec,
+        "logs_dir": logs_path,
+    }
+    try:
+        run_data = run_command_with_timeout(
+            command=cmd,
+            timeout=setup_entry.timeout_s,
+            stdouterr_path=stdouterr_path,
+            run_id=f"data_setup-{setup_entry.name}-{int(started_exec)}",
+            fancy=os.environ.get("CURATOR_BENCHMARKING_DEBUG", "0") == "0",
+        )
+        duration = time.time() - started_exec
+        success = run_data["returncode"] == 0 and not run_data["timed_out"]
+        result_data.update(
+            {
+                "exec_time_s": duration,
+                "exit_code": run_data["returncode"],
+                "timed_out": run_data["timed_out"],
+                "success": success,
+            }
+        )
+        if success:
+            logger.info(f"\tData setup {setup_entry.name} completed in {duration:.1f}s")
+        else:
+            logger.error(f"\tData setup {setup_entry.name} failed in {duration:.1f}s")
+            if run_data["timed_out"]:
+                logger.warning(f"\t⏰ Timed out after {setup_entry.timeout_s}s")
+            logger.error(f"\t➡️  Full output here: {stdouterr_path}")
+        return success
+    finally:
+        (setup_path / "results.json").write_text(json.dumps(get_obj_for_json(result_data)))
+
+
+def run_data_setups(
+    setup_entries: list[Entry],
+    path_resolver: PathResolver,
+    dataset_resolver: DatasetResolver,
+    session_path: Path,
+) -> bool:
+    """Run all configured data setup entries before benchmark execution."""
+    if not setup_entries:
+        return True
+
+    setup_root_path = session_path / "data_setup"
+    create_or_overwrite_dir(setup_root_path)
+    logger.info("Data setup entries to be run before benchmarks:")
+    for idx, setup_entry in enumerate(setup_entries, start=1):
+        logger.info(f"\t{idx}. {setup_entry.name}")
+
+    overall_success = True
+    for setup_entry in setup_entries:
+        overall_success &= run_data_setup_entry(
+            setup_entry=setup_entry,
+            path_resolver=path_resolver,
+            dataset_resolver=dataset_resolver,
+            setup_root_path=setup_root_path,
+        )
+        if not overall_success:
+            break
+    return overall_success
+
+
+def run_entry(  # noqa: PLR0913
+    entry: Entry,
+    path_resolver: PathResolver,
+    dataset_resolver: DatasetResolver,
+    session_entry_path: Path,
+    result_data: dict[str, Any],
+    gpu_stats_recorder_interval_s: float = 1.0,
+) -> bool:
+    # session_entry_path : This is the directory where benchmark results are stored
+    # scratch_path : This is for scratch/temp data; it is cleaned up if delete_scratch is True.
+    # ray_cluster_path : This is the directory where Ray debug/log files are saved
+    # logs_path : This is the directory where stdout/stderr and Ray startup logs are saved
+    scratch_path, ray_cluster_path, logs_path = [
+        (session_entry_path / d).absolute() for d in ["scratch", "ray_cluster", "logs"]
+    ]
+    cmd = entry.get_command_to_run(session_entry_path, path_resolver, dataset_resolver)
+    stdouterr_path = logs_path / "stdouterr.log"
+    run_id = result_data.get("run_id", f"{entry.name}-{int(time.time())}")
+    ray_client = ray_temp_dir = None
+    ray_num_cpus = entry.ray.get("num_cpus", os.cpu_count() or 1)
+    ray_num_gpus = entry.ray.get("num_gpus", 0)
+    ray_enable_object_spilling = bool(entry.ray.get("enable_object_spilling", False))
+
+    try:
+        # Create subdirs individually. logs_path uses ensure_dir (not create_or_overwrite_dir)
+        # to preserve any log output already written before run_entry() was called.
+        for directory in [scratch_path, ray_cluster_path]:
+            create_or_overwrite_dir(directory)
+        ensure_dir(logs_path)
+
+        ray_client, ray_temp_dir = setup_ray_cluster_and_env(
+            num_cpus=ray_num_cpus,
+            num_gpus=ray_num_gpus,
+            enable_object_spilling=ray_enable_object_spilling,
+            ray_log_path=logs_path / "ray.log",
+            object_store_size=None if entry.object_store_size == "default" else entry.object_store_size,
+        )
+
+        # Prepopulate <session_entry_path>/params.json with entry params.
+        # These will be appended with the benchmark params by the benchmark script.
+        (session_entry_path / "params.json").write_text(
+            json.dumps(
+                {
+                    "object_store_size_bytes": entry.object_store_size,
+                    "ray_num_cpus": ray_num_cpus,
+                    "ray_num_gpus": ray_num_gpus,
+                    "ray_enable_object_spilling": ray_enable_object_spilling,
+                    "entry_timeout_s": entry.timeout_s,
+                },
+                default=get_obj_for_json,
+                indent=2,
+            )
+        )
+
+        # Execute command with timeout, capturing GPU stats before and after
+        ray_cluster_data = get_ray_cluster_data()
+        gpu_stats_before = get_gpu_stats()
+        logger.info("\tGPU stats (before):")
+        warnings = log_gpu_stats(
+            gpu_stats_before,
+            warn_if_in_use=True,
+            warning_threshold=entry.gpu_mem_use_warning_threshold,
+            warning_threshold_msg="used before benchmark started",
+        )
+        logger.info(f"\tRunning command {' '.join(cmd) if isinstance(cmd, list) else cmd}")
+        # Background poller writes per-GPU stats (utilization, memory, temperature, processes) to
+        # gpustats.csv every gpu_stats_recorder_interval_s seconds. Set to 0 in YAML to disable.
+        gpu_stats_recorder_ctx = (
+            GPUStatsRecorder(
+                output_path=session_entry_path / "gpustats.csv",
+                interval_s=gpu_stats_recorder_interval_s,
+            )
+            if gpu_stats_recorder_interval_s > 0
+            else nullcontext()
+        )
+        started_exec = time.time()
+        with gpu_stats_recorder_ctx:
+            run_data = run_command_with_timeout(
+                command=cmd,
+                timeout=entry.timeout_s,
+                stdouterr_path=stdouterr_path,
+                run_id=run_id,
+                fancy=os.environ.get("CURATOR_BENCHMARKING_DEBUG", "0") == "0",
+            )
+        ended_exec = time.time()
+        logger.info("\tGPU stats (after):")
+        warnings += log_gpu_stats(
+            get_gpu_stats(),
+            warn_if_in_use=True,
+            warning_threshold=entry.gpu_mem_use_warning_threshold,
+            warning_threshold_msg="left in use after benchmark ended",
+        )
+        duration = ended_exec - started_exec
+
+        # Update result_data
+        result_data.update(
+            {
+                "cmd": cmd,
+                "exec_started_at": started_exec,
+                "exec_time_s": duration,
+                "exit_code": run_data["returncode"],
+                "timed_out": run_data["timed_out"],
+                "logs_dir": logs_path,
+                "ray_cluster_data": ray_cluster_data,
+                "gpu_stats": gpu_stats_before,
+                "warnings": warnings,
+            }
+        )
+        # script_persisted_data is a dictionary with keys "params" and "metrics"
+        # "params" will contain everything the script wrote to its params.json file
+        # "metrics" will contain everything the script wrote to its metrics.json file plus metrics
+        # from the Task objects restored from the tasks.pkl file.
+        script_persisted_data = get_entry_script_persisted_data(session_entry_path)
+        result_data.update(
+            {
+                "metrics": script_persisted_data["metrics"],
+                "params": script_persisted_data["params"],
+            }
+        )
+
+        # Check if the run itself returned a success code, if so, use the updated
+        # result_data to check if requirements were met.
+        if run_data["returncode"] == 0:
+            success = check_requirements_update_results(result_data, entry.requirements)
+        else:
+            success = False
+            logger.error(f"\t❌ Run Failed in {duration:.1f} seconds")
+            if run_data["timed_out"]:
+                logger.warning(f"\t⏰ Timed out after {entry.timeout_s}s")
+            logger.error(f"\t➡️  Full output here: {stdouterr_path}")
+
+        result_data["success"] = success
+        logger.info(f"\tLogs found in {logs_path}")
+        Path(session_entry_path / "results.json").write_text(json.dumps(get_obj_for_json(result_data)))
+
+        return success
+
+    finally:
+        teardown_ray_cluster_and_env(ray_client, ray_temp_dir, ray_cluster_path)
+
+        # Clean up the scratch dir if configured to delete
+        if entry.delete_scratch:
+            shutil.rmtree(scratch_path, ignore_errors=True)
+
+
+def run_benchmarks(args: Namespace) -> int:  # noqa: C901, PLR0911, PLR0912, PLR0915
+    if args.path_mode:
+        set_path_mode(args.path_mode)
+
+    # Consolidate the configuration from all YAML files into a single dict
+    config_dict = merge_config_files(args.config)
+
+    # Preprocess the config dict prior to creating objects from it
+    try:
+        assert_valid_config_dict(config_dict)
+        config_dict = remove_disabled_blocks(config_dict)
+        config_dict = resolve_env_vars(config_dict, strict=args.strict_config_check)
+    except ValueError as e:
+        logger.error(f"Invalid configuration: {e}")
+        return 1
+
+    if args.entries is not None and args.entries_exact is not None:
+        logger.error("--entries and --entries-exact are mutually exclusive")
+        return 1
+
+    entries_exact_list: list[str] | None = None
+    if args.entries_exact is not None:
+        entries_exact_list = [name.strip() for name in args.entries_exact.split(",") if name.strip()]
+        if not entries_exact_list:
+            logger.error("--entries-exact must contain at least one non-empty name")
+            return 1
+
+    # Now that all YAML config files have been read, merged, and processed, create the Session object.
+    try:
+        session = Session.from_dict(
+            config_dict,
+            entry_filter_expr=args.entries,
+            entries_exact=entries_exact_list,
+        )
+    except (TypeError, ValueError) as e:
+        logger.error(str(e))
+        return 1
+
+    # GPU stats recorder config: polls every interval_s seconds while each entry runs.
+    # Default 1.0 (1 Hz). Set to 0 to disable.
+    gpu_stats_recorder_interval_s = float(config_dict.get("gpu_stats_recorder", {}).get("interval_s", 1.0))
+
+    if args.list:
+        for entry in session.entries:
+            logger.info(f"\t{entry.name}")
+        return 0
+
+    # Create session folder under results_dir
+    session_name = args.session_name or time.strftime("benchmark-run__%Y-%m-%d_%H-%M-%S_UTC")
+    session_path = (session.results_path / session_name).absolute()
+    ensure_dir(session_path)
+
+    if args.reason:
+        session.run_reason = args.reason
+    if args.viewer_url:
+        session.viewer_url = args.viewer_url
+        session.viewer_url_template = None
+    elif args.viewer_url_template:
+        session.viewer_url = None
+        session.viewer_url_template = args.viewer_url_template
+    try:
+        # viewer_url must have paths that are visible outside of a
+        # container, meaning host paths that are mapped to container
+        # mounts (if any) will be "unmapped".
+        session.viewer_url = resolve_viewer_url(
+            viewer_url=session.viewer_url,
+            viewer_url_template=session.viewer_url_template,
+            results_path=session.path_resolver.unmap_container_path(session.results_path),
+            session_name=session_name,
+            session_path=session.path_resolver.unmap_container_path(session_path),
+        )
+    except ValueError as e:
+        logger.error(str(e))
+        return 1
+
+    session_overall_success = True
+    logger.info(f"Started session {session_name}...")
+    env_dict = dump_env(session_obj=session, output_path=session_path)
+
+    if not run_data_setups(
+        setup_entries=session.data_setups,
+        path_resolver=session.path_resolver,
+        dataset_resolver=session.dataset_resolver,
+        session_path=session_path,
+    ):
+        logger.error("Data setup failed; benchmark entries will not be run.")
+        return 1
+
+    active_sinks = initialize_sinks(session.sinks, session_name=session_name, session=session, env_dict=env_dict)
+
+    # Print a summary of the entries that will be run in the for loop below
+    # Disabled entries will not be printed
+    logger.info("Benchmark entries to be run in this session:")
+    for idx, entry in enumerate(session.entries, start=1):
+        logger.info(f"\t{idx}. {entry.name}")
+
+    for entry in session.entries:
+        run_success = False
+        run_id = f"{entry.name}-{int(time.time())}"
+        result_data = {
+            "name": entry.name,
+            "run_id": run_id,
+            "success": run_success,
+        }
+        # Derive the stdouterr log path and add it as a loguru sink so all log
+        # output for this entry (including pre-subprocess errors) is captured.
+        session_entry_path = (session_path / entry.name).absolute()
+        entry_logs_path = session_entry_path / "logs"
+        entry_stdouterr_path = entry_logs_path / "stdouterr.log"
+        entry_logs_path.mkdir(parents=True, exist_ok=True)
+        entry_log_id = logger.add(entry_stdouterr_path, mode="a", colorize=False)
+        logger.info(f"🚀 Running {entry.name} (run ID: {run_id})")
+
+        for sink in active_sinks:
+            call_sink_hook(
+                sink,
+                "register_benchmark_entry_starting",
+                context=f"registering {entry.name} as starting",
+                result_dict=result_data,
+                benchmark_entry=entry,
+            )
+
+        try:
+            run_success = run_entry(
+                entry=entry,
+                path_resolver=session.path_resolver,
+                dataset_resolver=session.dataset_resolver,
+                session_entry_path=session_entry_path,
+                result_data=result_data,
+                gpu_stats_recorder_interval_s=gpu_stats_recorder_interval_s,
+            )
+
+        except Exception as e:
+            run_success = False
+            error_traceback = traceback.format_exc()
+            logger.exception("\t\t❌ Entry failed with exception")
+            result_data.update(
+                {
+                    "error": str(e),
+                    "traceback": error_traceback,
+                    "success": run_success,
+                }
+            )
+
+        finally:
+            session_overall_success &= run_success
+            for sink in active_sinks:
+                call_sink_hook(
+                    sink,
+                    "register_benchmark_entry_finished",
+                    context=f"registering {entry.name} as finished",
+                    result_dict=result_data,
+                    benchmark_entry=entry,
+                )
+            logger.remove(entry_log_id)
+
+    for sink in active_sinks:
+        call_sink_hook(sink, "finalize", context="finalizing benchmark reporting sinks")
+    logger.info(f"Session {session_name} completed with overall success: {session_overall_success}")
+    return 0 if session_overall_success else 1
