+++ benchmarking/runner/path_resolver.py
@@ -12,16 +12,52 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
+import os
 from collections.abc import Iterator
 from pathlib import Path
 
 CONTAINER_CURATOR_DIR = "/opt/Curator"
+CURATOR_BENCHMARK_PATH_MODE_ENV = "CURATOR_BENCHMARK_PATH_MODE"
 
 # Prefix prepended to host paths when resolving container paths that do not have an explicit
 # container_path specified. Using /MOUNT makes it obvious in log/error messages from the
 # container that these paths are available on the host, and allows copy-and-paste of all
 # but the "/MOUNT" prefix to get the equivalent host path.
 DEFAULT_CONTAINER_PATH_PREFIX = "/MOUNT"
+VALID_PATH_MODES = {"auto", "host", "container"}
+
+
+def set_path_mode(mode: str) -> None:
+    """Set the process-wide path-resolution mode used by PathResolver."""
+    mode = mode.strip().lower()
+    if mode not in VALID_PATH_MODES:
+        msg = f"{CURATOR_BENCHMARK_PATH_MODE_ENV} must be one of {sorted(VALID_PATH_MODES)}; got {mode!r}"
+        raise ValueError(msg)
+    os.environ[CURATOR_BENCHMARK_PATH_MODE_ENV] = mode
+
+
+def get_path_mode() -> str:
+    """Return the requested path-resolution mode."""
+    mode = os.environ.get(CURATOR_BENCHMARK_PATH_MODE_ENV, "auto").strip().lower() or "auto"
+    if mode not in VALID_PATH_MODES:
+        msg = f"{CURATOR_BENCHMARK_PATH_MODE_ENV} must be one of {sorted(VALID_PATH_MODES)}; got {mode!r}"
+        raise ValueError(msg)
+    return mode
+
+
+def _auto_detect_container() -> bool:
+    """Best-effort compatibility fallback for callers that do not set path mode."""
+    return Path("/.dockerenv").exists() or Path("/run/.containerenv").exists()
+
+
+def use_container_paths() -> bool:
+    """Return whether configured path names should resolve to container paths."""
+    mode = get_path_mode()
+    if mode == "container":
+        return True
+    if mode == "host":
+        return False
+    return _auto_detect_container()
 
 
 class PathResolver:
@@ -38,8 +74,7 @@ class PathResolver:
         'container_path' overrides the default container path (which is the host_path
         prefixed with '/MOUNT').
         """
-        # TODO: Is this the best way to determine if running inside a Docker container?
-        in_docker = Path("/.dockerenv").exists()
+        use_container_path = use_container_paths()
         self.path_map: dict[str, Path] = {}
         self._volume_pairs: list[tuple[Path, Path]] = []
 
@@ -51,7 +86,7 @@ class PathResolver:
                 container_path = (
                     Path(raw_container) if raw_container else Path(f"{DEFAULT_CONTAINER_PATH_PREFIX}/{host_path}")
                 )
-                self.path_map[name] = container_path if in_docker else host_path
+                self.path_map[name] = container_path if use_container_path else host_path
                 self._volume_pairs.append((host_path, container_path))
         else:
             # Legacy top-level YAML path fields are deprecated. The path
@@ -63,7 +98,7 @@ class PathResolver:
                 ("model_weights_path", Path(data["model_weights_path"])),
             ]:
                 container_path = Path(f"{DEFAULT_CONTAINER_PATH_PREFIX}/{host_path}")
-                self.path_map[name] = container_path if in_docker else host_path
+                self.path_map[name] = container_path if use_container_path else host_path
                 self._volume_pairs.append((host_path, container_path))
 
     def volume_mount_pairs(self) -> Iterator[tuple[Path, Path]]:
